# openapi.swagger.yaml
swagger: "2.0"
info:
  title: VSFV API
  version: v1
  contact:
    name: pixlcrashr
    url: https://pixlcrashr.dev
    email: info@pixlcrashr.dev
...
host: test.pixlcrashr.dev
basePath: /api
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/organizations:
    get:
      summary: |-
        Lists organizations with pagination.
        Authorization:
          Scope: organizations:read
          Permission: organizations:read
          Domain: global
      operationId: OrganizationService_ListOrganizations
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1ListOrganizationsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: page_size
          description: |-
            Maximum number of organizations to return. The service may return fewer.
            If unspecified, at most 20 are returned. Maximum value is 100.
          in: query
          required: false
          type: integer
          format: int32
        - name: page_token
          description: A page token from a previous ListOrganizations call.
          in: query
          required: false
          type: string
        - name: filter
          description: |-
            Filter expression conforming to AIP-160.
            Supported fields: display_name.
            Example: "display_name=\"Acme\"".
          in: query
          required: false
          type: string
        - name: order_by
          description: Order by expression (e.g. "display_name", "create_time desc").
          in: query
          required: false
          type: string
      tags:
        - OrganizationService
    post:
      summary: |-
        Creates a new organization.
        Authorization:
          Scope: organizations:write
          Permission: organizations:create
          Domain: global
      operationId: OrganizationService_CreateOrganization
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1Organization'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: organization
          description: The organization to create.
          in: body
          required: true
          schema:
            $ref: '#/definitions/v1Organization'
        - name: organization_id
          description: |-
            The ID to use for the organization. If not provided, a system-generated
            UUID will be used. Must be unique across all organizations.
          in: query
          required: false
          type: string
      tags:
        - OrganizationService
...
definitions:
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
...
  v1ListOrganizationsResponse:
    type: object
    properties:
      organizations:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1Organization'
        description: The organizations returned.
      next_page_token:
        type: string
        description: A token to retrieve the next page of results.
      total_size:
        type: string
        format: int64
        description: Total number of organizations matching the filter (may be an estimate).
...
  v1Month:
    type: string
    enum:
      - MONTH_UNSPECIFIED
      - MONTH_JANUARY
      - MONTH_FEBRUARY
      - MONTH_MARCH
      - MONTH_APRIL
      - MONTH_MAY
      - MONTH_JUNE
      - MONTH_JULY
      - MONTH_AUGUST
      - MONTH_SEPTEMBER
      - MONTH_OCTOBER
      - MONTH_NOVEMBER
      - MONTH_DECEMBER
    default: MONTH_UNSPECIFIED
...
  v1Organization:
    type: object
    properties:
      name:
        type: string
        title: |-
          The resource name of the organization.
          Format: organizations/{organization}
        readOnly: true
      uid:
        type: string
        description: The UUID of the organization.
        readOnly: true
      display_name:
        type: string
        description: Human-readable organization name.
      display_description:
        type: string
        description: Optional free-text description.
      start_month:
        $ref: '#/definitions/v1Month'
        description: The start month of ledger years for this organization. It is assumed that the start date is the 1st of the month.
      update_time:
        type: string
        format: date-time
        description: Last modification timestamp.
        readOnly: true
      create_time:
        type: string
        format: date-time
        description: Creation timestamp.
        readOnly: true
      etag:
        type: string
        description: Entity tag for optimistic concurrency control.
        readOnly: true
    description: Organization is the top-level resource that owns all other resources.
    required:
      - display_name
      - start_month
...